Frontend: PetSlice Implemented

+++ frontend/src/components/admin/AdminProfile.tsx
@@ -5,16 +5,18 @@ import { User } from "@/interfaces/user";
 import axios from "axios";
 import Link from "next/link";
 import React, { useEffect, useState } from "react";
-import UserCard from "../pet/UserCard";
 import { Vet } from "@/interfaces/vet";
 import { Trainer } from "@/interfaces/trainer";
 import { Pet } from "@/interfaces/pet";
 import { PetCard } from "../pet/PetCard";
 import axiosInstance from "@/utils/axios";
 import toast from "react-hot-toast";
+import { TrainerCard } from "../trainer/TrainerCard";
+import { UserCard } from "../user/UserCard";
+import { VetCard } from "../vet/VetCard";
 
 function AdminProfile() {
-  const [activeTab, setActiveTab] = useState("Profile");
+  const [activeTab, setActiveTab] = useState("ServicesBookings");
   const handleTabClick = (tab: string) => {
     setActiveTab(tab);
   };
@@ -28,45 +30,67 @@ function AdminProfile() {
   const [trainings, setTrainings] = useState<Training[]>([]);
 
   const handleUserDelete = async (userId: string) => {
-    try {
-      const response = await axios.delete(`${process.env.HOST}/user/${userId}`);
-      toast.success("User is Deleted")
-      setUsers(users.filter((user) => user._id !== userId));
-    } catch (error) {
-      console.error(error);
+    const isConfirmed = window.confirm(
+      `Are you sure you want to delete this user?`
+    );
+    if (isConfirmed) {
+      try {
+        const response = await axios.delete(
+          `${process.env.HOST}/user/delete/${userId}`
+        );
+        toast.success("User is Deleted");
+        setUsers(users.filter((user) => user._id !== userId));
+      } catch (error) {
+        console.error(error);
+      }
     }
   };
 
   const handleVetDelete = async (vetId: string) => {
-    try {
-      const response = await axios.delete(`${process.env.HOST}/vet/${vetId}`);
-  
-      toast.success("Vet is Deleted")
-      setVets(vets.filter((vet) => vet._id !== vetId));
-    } catch (error) {
-      console.error(error);
+    const isConfirmed = window.confirm(
+      `Are you sure you want to delete this vet?`
+    );
+    if (isConfirmed) {
+      try {
+        const response = await axios.delete(`${process.env.HOST}/vet/${vetId}`);
+
+        toast.success("Vet is Deleted");
+        setVets(vets.filter((vet) => vet._id !== vetId));
+      } catch (error) {
+        console.error(error);
+      }
     }
   };
 
   const handleTrainerDelete = async (trainerId: string) => {
-    try {
-      const response = await axios.delete(
-        `${process.env.HOST}/trainer/${trainerId}`
-      );
-      toast.success("Trainer is Deleted")
-      setTrainers(trainers.filter((trainer) => trainer._id !== trainerId));
-    } catch (error) {
-      console.error(error);
+    const isConfirmed = window.confirm(
+      `Are you sure you want to delete this trainer?`
+    );
+    if (isConfirmed) {
+      try {
+        const response = await axios.delete(
+          `${process.env.HOST}/trainer/${trainerId}`
+        );
+        toast.success("Trainer is Deleted");
+        setTrainers(trainers.filter((trainer) => trainer._id !== trainerId));
+      } catch (error) {
+        console.error(error);
+      }
     }
   };
 
   const handlePetDelete = async (petId: string) => {
-    try {
-      const response = await axios.delete(`${process.env.HOST}/pet/${petId}`);
-      toast.success("Pet is Deleted")
-      setPets(pets.filter((pet) => pet._id !== petId));
-    } catch (error) {
-      console.error(error);
+    const isConfirmed = window.confirm(
+      `Are you sure you want to delete this user?`
+    );
+    if (isConfirmed) {
+      try {
+        const response = await axios.delete(`${process.env.HOST}/pet/${petId}`);
+        toast.success("Pet is Deleted");
+        setPets(pets.filter((pet) => pet._id !== petId));
+      } catch (error) {
+        console.error(error);
+      }
     }
   };
 
@@ -113,8 +137,11 @@ function AdminProfile() {
         return (
           <div>
             <div className="relative overflow-x-auto shadow-md sm:rounded-lg">
-              <table className="w-full text-sm text-left rtl:text-right text-gray-500 dark:text-gray-400">
-                <thead className="text-xs text-gray-700 uppercase bg-gray-50 dark:bg-gray-700 dark:text-gray-400">
+              <table className="w-full text-dark-blue ">
+                <thead
+                  style={{ fontSize: "16px" }}
+                  className=" text-red-200 bg-dark-blue"
+                >
                   <tr>
                     <th scope="col" className="px-6 py-3">
                       Client name
@@ -148,37 +175,37 @@ function AdminProfile() {
                       key={index}
                       className={`${
                         index % 2 === 0
-                          ? "odd:bg-white odd:dark:bg-gray-900"
-                          : "even:bg-gray-50 even:dark:bg-gray-800"
-                      } border-b dark:border-gray-700`}
+                          ? "odd:bg-gray-200 "
+                          : "even:bg-gray-100"
+                      } border-b `}
                     >
-                      <td className="px-6 py-4 font-medium text-gray-900 whitespace-nowrap dark:text-white">
+                      <td className="px-6 py-4 font-bold text-gray-700 whitespace-nowrap ">
                         {service.user_name}
                       </td>
                       <td className="px-6 py-4">{service.pet_species}</td>
-                      <td className="px-6 py-4">{service.totalPrice}</td>
+                      <td className="px-6 py-4 ">{service.totalPrice}</td>
                       <td className="px-6 py-4">{service.city}</td>
                       <td className="px-6 py-4">{service.booking_date}</td>
                       <td className="px-6 py-4">{service.booking_time}</td>
                       <td className="px-6 py-4">
                         {service.isConfirmed ? (
-                          <button className="bg-green-500 text-white px-3 py-1 rounded-md mr-2">
+                          <button className="bg-green-500 text-white px-3 py-2 rounded-md mr-2 w-56 font-bold">
                             Paid
                           </button>
                         ) : (
-                          <button className="bg-red-500 text-white px-3 py-1 rounded-md mr-2">
+                          <button className="bg-red-500 text-white px-3 py-2 rounded-md mr-2 w-56 font-bold">
                             Not Paid
                           </button>
                         )}
                       </td>
                       <td className="px-6 py-4">
                         {service.vetId ? (
-                          <span className="bg-saddle-brown text-white px-3 py-1 rounded-md mr-2">
+                          <span className="bg-saddle-brown text-white px-3 py-2 rounded-md mr-2 font-bold">
                             Assigned
                           </span>
                         ) : (
                           <Link
-                            className="bg-blue-600 text-white px-3 py-1 rounded-md mr-2 no-underline"
+                            className="bg-blue-600 text-white px-3 py-2 rounded-md mr-2 no-underline font-bold"
                             href={`/assignVet?bookingId=${service._id}`}
                           >
                             Assign Vet
@@ -196,8 +223,11 @@ function AdminProfile() {
         return (
           <div>
             <div className="relative overflow-x-auto shadow-md sm:rounded-lg">
-              <table className="w-full text-sm text-left rtl:text-right text-gray-500 dark:text-gray-400">
-                <thead className="text-xs text-gray-700 uppercase bg-gray-50 dark:bg-gray-700 dark:text-gray-400">
+              <table className="w-full text-dark-blue  ">
+                <thead
+                  style={{ fontSize: "16px" }}
+                  className=" text-red-200 bg-dark-blue"
+                >
                   <tr>
                     <th scope="col" className="px-6 py-3">
                       Client name
@@ -231,11 +261,11 @@ function AdminProfile() {
                       key={index}
                       className={`${
                         index % 2 === 0
-                          ? "odd:bg-white odd:dark:bg-gray-900"
-                          : "even:bg-gray-50 even:dark:bg-gray-800"
-                      } border-b dark:border-gray-700`}
+                          ? "odd:bg-gray-200 "
+                          : "even:bg-gray-100"
+                      } border-b `}
                     >
-                      <td className="px-6 py-4 font-medium text-gray-900 whitespace-nowrap dark:text-white">
+                      <td className="px-6 py-4 font-bold text-gray-700 whitespace-nowrap">
                         {training.user_name}
                       </td>
                       <td className="px-6 py-4">{training.pet_species}</td>
@@ -245,23 +275,23 @@ function AdminProfile() {
                       <td className="px-6 py-4">{training.booking_time}</td>
                       <td className="px-6 py-4">
                         {training.isConfirmed ? (
-                          <button className="bg-green-500 text-white px-3 py-1 rounded-md mr-2">
+                          <button className="bg-green-500 text-white px-3 py-1 rounded-md mr-2 w-56 font-bold">
                             Paid
                           </button>
                         ) : (
-                          <button className="bg-red-500 text-white px-3 py-1 rounded-md mr-2">
+                          <button className="bg-red-500 text-white px-3 py-1 rounded-md mr-2 w-56 font-bold">
                             Not Paid
                           </button>
                         )}
                       </td>
                       <td className="px-6 py-4">
                         {training.trainerId ? (
-                          <span className="bg-saddle-brown text-white px-3 py-1 rounded-md mr-2">
+                          <span className="bg-saddle-brown text-white px-3 py-1 rounded-md mr-2 font-bold">
                             Assigned
                           </span>
                         ) : (
                           <Link
-                            className="bg-blue-600 text-white px-3 py-1 rounded-md mr-2 no-underline"
+                            className="bg-blue-600 text-white px-3 py-1 rounded-md mr-2 no-underline font-bold"
                             href={`/assignTrainer?bookingId=${training._id}`}
                           >
                             Assign Trainer
@@ -278,27 +308,35 @@ function AdminProfile() {
 
       case "Users":
         return (
-          <div>
-            <div className="grid grid-cols-1 gap-4 sm:grid-cols-2 lg:grid-cols-3">
-              {users.map((user) => (
+          <div className="row ">
+            {users.map((user) => (
+              <div className="col-md-4">
                 <div key={Math.random()}>
-                  <UserCard user={user} />
-                  <button onClick={() => handleUserDelete(user._id)}>
+                  <div key={user._id}>
+                    <UserCard user={user} />
+                  </div>
+                  <button
+                    onClick={() => handleUserDelete(user._id)}
+                    className="mt-2  px-4 py-2 bg-red-500 text-white rounded-md"
+                  >
                     Delete
                   </button>
                 </div>
-              ))}
-            </div>
+              </div>
+            ))}
           </div>
         );
-      case "Trainers":
+      case "Vets":
         return (
-          <div>
-            <div className="grid grid-cols-1 gap-4 sm:grid-cols-2 lg:grid-cols-3">
+          <div className="container-fluid mt-3">
+            <div className="row">
               {vets.map((vet) => (
                 <div key={Math.random()}>
-                  <UserCard user={vet} />
-                  <button onClick={() => handleVetDelete(vet._id)}>
+                  <VetCard user={vet} />
+                  <button
+                    onClick={() => handleVetDelete(vet._id)}
+                    className="mt-2 px-4 py-2 bg-red-500 text-white rounded-md"
+                  >
                     Delete
                   </button>
                 </div>
@@ -307,18 +345,21 @@ function AdminProfile() {
           </div>
         );
 
-      case "Vets":
+      case "Trainers":
         return (
-          <div>
-            <div className="grid grid-cols-1 gap-4 sm:grid-cols-2 lg:grid-cols-3">
-              {trainers.map((trainer) => (
-                <div key={Math.random()}>
-                  <UserCard user={trainer} />
-                  <button onClick={() => handleTrainerDelete(trainer._id)}>
-                    Delete
-                  </button>
-                </div>
-              ))}
+          <div className="container-fluid mt-3">
+            <div className="row">
+                {trainers.map((trainer) => (
+                  <div key={Math.random()}>
+                    <TrainerCard user={trainer} />
+                    <button
+                      onClick={() => handleTrainerDelete(trainer._id)}
+                      className="mt-2 px-4 py-2 bg-red-500 text-white rounded-md"
+                    >
+                      Delete
+                    </button>
+                  </div>
+                ))}
             </div>
           </div>
         );
@@ -330,9 +371,14 @@ function AdminProfile() {
               {pets.map((pet) => (
                 <div key={Math.random()}>
                   <PetCard pet={pet} />
-                  <button onClick={() => handlePetDelete(pet._id)}>
-                    Delete
-                  </button>
+                  <div className="mx-6">
+                    <button
+                      onClick={() => handlePetDelete(pet._id)}
+                      className="mt-2 px-4 py-2 bg-red-500 text-white rounded-md"
+                    >
+                      Delete
+                    </button>
+                  </div>
                 </div>
               ))}
             </div>
@@ -345,16 +391,17 @@ function AdminProfile() {
 
   return (
     <div>
-      <div className=" p-9 bg-white border border-gray-200 rounded-lg shadow m-8 border-1">
-        <div className="text-sm font-medium text-center text-gray-500 border-b border-gray-200 dark:text-gray-400 dark:border-gray-700">
+      <div className=" p-9 ">
+        <div className="text-sm font-medium text-center text-gray-500 ">
           <ul className="flex flex-wrap -mb-px">
             <li className="me-2">
               <button
                 onClick={() => handleTabClick("ServicesBookings")}
+                style={{ fontSize: "18px" }}
                 className={`inline-block p-4 border-b-2 rounded-t-lg ${
-                  activeTab === "profile"
-                    ? "border-blue-600 text-blue-600 dark:border-blue-500 dark:text-blue-500"
-                    : "border-transparent text-gray-600 hover:text-gray-600 hover:border-gray-300 dark:hover:text-gray-300"
+                  activeTab === "ServicesBookings"
+                    ? "text-saddle-brown font-bold font-2xl border-saddle-brown"
+                    : " border-transparent text-dark-blue hover:text-saddle-brown"
                 }`}
               >
                 Services Bookings
@@ -363,10 +410,11 @@ function AdminProfile() {
             <li className="me-2">
               <button
                 onClick={() => handleTabClick("TrainingBookings")}
+                style={{ fontSize: "18px" }}
                 className={`inline-block p-4 border-b-2 rounded-t-lg ${
-                  activeTab === "dashboard"
-                    ? "border-blue-600 text-blue-600 dark:border-blue-500 dark:text-blue-500"
-                    : "border-transparent text-gray-600 hover:text-gray-600 hover:border-gray-300 dark:hover:text-gray-300"
+                  activeTab === "TrainingBookings"
+                    ? "text-saddle-brown font-bold font-2xl border-saddle-brown"
+                    : " border-transparent text-dark-blue hover:text-saddle-brown"
                 }`}
               >
                 Training Bookings
@@ -375,10 +423,11 @@ function AdminProfile() {
             <li className="me-2">
               <button
                 onClick={() => handleTabClick("Users")}
+                style={{ fontSize: "18px" }}
                 className={`inline-block p-4 border-b-2 rounded-t-lg ${
-                  activeTab === "settings"
-                    ? "border-blue-600 text-blue-600 dark:border-blue-500 dark:text-blue-500"
-                    : "border-transparent text-gray-600 hover:text-gray-600 hover:border-gray-300 dark:hover:text-gray-300"
+                  activeTab === "Users"
+                    ? "text-saddle-brown font-bold font-2xl border-saddle-brown"
+                    : " border-transparent text-dark-blue hover:text-saddle-brown"
                 }`}
               >
                 Users
@@ -387,10 +436,11 @@ function AdminProfile() {
             <li className="me-2">
               <button
                 onClick={() => handleTabClick("Trainers")}
+                style={{ fontSize: "18px" }}
                 className={`inline-block p-4 border-b-2 rounded-t-lg ${
-                  activeTab === "settings"
-                    ? "border-blue-600 text-blue-600 dark:border-blue-500 dark:text-blue-500"
-                    : "border-transparent text-gray-600 hover:text-gray-600 hover:border-gray-300 dark:hover:text-gray-300"
+                  activeTab === "Trainers"
+                    ? "text-saddle-brown font-bold font-2xl border-saddle-brown"
+                    : " border-transparent text-dark-blue hover:text-saddle-brown"
                 }`}
               >
                 Trainers
@@ -399,10 +449,11 @@ function AdminProfile() {
             <li className="me-2">
               <button
                 onClick={() => handleTabClick("Vets")}
+                style={{ fontSize: "18px" }}
                 className={`inline-block p-4 border-b-2 rounded-t-lg ${
-                  activeTab === "settings"
-                    ? "border-blue-600 text-blue-600 dark:border-blue-500 dark:text-blue-500"
-                    : "border-transparent text-gray-600 hover:text-gray-600 hover:border-gray-300 dark:hover:text-gray-300"
+                  activeTab === "Vets"
+                    ? "text-saddle-brown font-bold font-2xl border-saddle-brown"
+                    : " border-transparent text-dark-blue hover:text-saddle-brown"
                 }`}
               >
                 Vets
@@ -411,10 +462,11 @@ function AdminProfile() {
             <li className="me-2">
               <button
                 onClick={() => handleTabClick("Pets")}
+                style={{ fontSize: "18px" }}
                 className={`inline-block p-4 border-b-2 rounded-t-lg ${
-                  activeTab === "contacts"
-                    ? "border-blue-600 text-blue-600 dark:border-blue-500 dark:text-blue-500"
-                    : "border-transparent text-gray-600 hover:text-gray-600 hover:border-gray-300 dark:hover:text-gray-300"
+                  activeTab === "Pets "
+                    ? "text-saddle-brown font-bold font-2xl border-saddle-brown"
+                    : " border-transparent text-dark-blue hover:text-saddle-brown"
                 }`}
               >
                 Pets
